Starter Commit
First commit
Every post on sethbotdeveloper channel will be "tracked" for 10 seconds.
after that 10 seconds, the number of posts on that specific post will be
written... alot more to do... but that was just me playing around today
;)

Made a basic "user class"
Added basic file input to save a json (Reading is still todo)
Added a collection reader.